requirements.txt
fair-esm
scikit-learn
biopython
matplotlib
torch
pandas
tqdm